requirements.txt
Django==1.7.6
mysql-python==1.2.5
djangorestframework
markdown
django-filter
django_compressor